CTX-3890: Changed width and height variables from float to int

+++ coretex/coretex/conversion/converters/pascal/instance_extractor.py
@@ -31,7 +31,7 @@ from ....annotation import CoretexSegmentationInstance, BBox
 from ....dataset import ImageDataset
 
 
-ContourPoints = List[List[float]]
+ContourPoints = List[List[int]]
 SegmentationPolygon = List[ContourPoints]
